_PM_stop: Don't HardFault when _PM_begin failed
When _PM_begin fails, it leaves the object in a partially-uninitialized
state.  To recover any allocated storage, we need to call _PM_free;
but if the allocation of core->screenData failed, then other internal
pointers such as core->oe.setReg are not set, and the actions in
_PM_stop will cause a HardFault.

This is a partial fix for adafruit/circuitpython#3184.